Removed clicked event from purchase button and changed to change event for facility.js
