+++ scripts/PurchaseBtn.js
@@ -9,17 +9,3 @@ export const purchaseBtn = () => {
     </div>
     `;
 };
-
-document.addEventListener('click', async (e) => {
-  const para = document.querySelector('.cartItems');
-  const copyOfTransientState = transientStateCopy();
-  const transientMineralId = copyOfTransientState.mineralId;
-
-  const response = await fetch(' http://localhost:8088/minerals');
-  const minerals = await response.json();
-  const matchedMineral = minerals.find(
-    (mineral) => mineral.id === transientMineralId
-  );
-
-  para.innerHTML = `1 ton of ${matchedMineral.name}`;
-});
